add sample to samples folder
